Implement a TypedCommand component
This component displays on the loading screen as a user typing.

Implementing this required various changes to the code, including
updating versions of the testing library so that the waitFor
function was available for the asynchronous tests required.

Implementing the useEffect hook on the component was particularly
tricky, and it's important that it returns a function. It's this
function that clears up after the hook, which removes a warning
that was being caused by setPosition being called while the
component was unmounted.