Rework flow for external links and contact
The ContactMe component was removed. Now, clicking contact in the
nav bar opens up an email with the correct email address.

The links in the ContactMe component are now found in the AboutMe
component, which makes more sense from a user perspective, as they
had nothing to do with making contact.