Add Main component
This was added to allow the use of different types of Router in
tests compared to the app itself. This fixes a problem with
BrowserRouter where the browser history persists between tests,
meaning that any tests that update or rely on the history can
fail depending on the order they are run in.